add mcu_powersave for other distros
